Перенесет логику воспроизведения в 'WithVideoPlayer'.
